edited dashboard

+++ database/migrations/2021_01_07_000011_add_relationship_fields_to_twitter_user_jobs_table.php
@@ -11,7 +11,7 @@ class AddRelationshipFieldsToTwitterUserJobsTable extends Migration
         Schema::table('twitter_user_jobs', function (Blueprint $table) {
             $table->unsignedBigInteger('user_id')->nullable();
             $table->foreign('user_id', 'user_fk_2933067')->references('id')->on('twitter_users');
-            $table->unsignedBigInteger('job_id');
+            $table->integer('job_id');
             $table->foreign('job_id', 'job_fk_2933071')->references('id')->on('jobs');
         });
     }
